[FEAT] PDF 시험지 업로드 및 자동 데이터 파싱(추출) 기능 구현 v3.2 - 정답지 수정 로직 구현 - 폴더를 삭제하는 경우 시험지/정답지도 같이 삭제되도록 수정 - @Transactional 설정 추가

+++ src/main/webapp/resources/js/admin_main.js
@@ -49,6 +49,9 @@ document.addEventListener('DOMContentLoaded', () => {
         // 폴더 삭제 버튼
         const folderDeleteBtn = e.target.closest('.btn-delete')
         if(folderDeleteBtn){
+            const confirmed = confirm('폴더를 삭제하시겠습니까? 해당 폴더에 포함된 시험지들도 함께 삭제됩니다.')
+            if(!confirmed) return
+            
             const folderId = folderDeleteBtn.getAttribute('data-id')
             deleteFolder(folderId)
         }
